refactor: colocate profile service tests
